requirements.txt
bcrypt==4.1.2
fastapi==0.110.0
strawberry-graphql==0.220.0
uvicorn[standard]==0.28.0
alembic==1.13.1
psycopg2-binary==2.9.9
PyJWT==2.8.0
asyncpg==0.29.0
SQLAlchemy[asyncio]==2.0.28
PyYAML==6.0.1
aiohttp==3.9.3
pydantic==2.7.4